requirements.txt
streamlit==1.28.1
pandas==2.0.3
pytesseract==0.3.10
Pillow==10.0.1
openpyxl==3.1.2